refactor: re-organize packages
Some minor functionality was removed to clean-up the API along the way.
That said the purpose of this commit to shift the code around a bit to
improve the organization of the various extension methods and make them
easier to use.

Specific extension methods can be found under packages with the pattern
`scala_result.extensions.*._`. While all `scala_result` extensions can
be imported using `scala_result.implicits._`.